requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pwdlib[argon2,bcrypt]
python-jose[cryptography]
python-multipart